requirements.txt
Django==1.8.6
Pillow==3.0.0
django-picklefield==0.3.2
wsgiref==0.1.2
